Fix broken imports after moving code
