+++ ldb/formTemplates/student.py
@@ -3,6 +3,8 @@ from __future__ import unicode_literals
 
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
+from django.utils.translation import ugettext_lazy as _
+
 
 from dienst2.extras import CharNullField
 from ldb.formTemplates.person import Person
